Add drop = FALSE when sampling from block in sa_pairing_generalized.R
Subsetting a single row, sample_block, from a matrix, block, yields a vector because R simplifies the object. However, cal_block_energy requires that the number of rows of sample_block can be evaulated. If sample_block is a vector instead, nrow will evaluate to NULL causing the function to fail. I added the drop = FALSE argument to make sure that sample_block remains a matrix when only one block is extracted.